Don't delete database table if it doesn't exist

+++ src/main/java/controllers/database/Database.java
@@ -45,7 +45,9 @@ public class Database {
     public static void delete(String tableName) {
         Path path = getTablePath(tableName);
         try {
-            Files.delete(path);
+            if (Files.exists(path)) {
+                Files.delete(path);
+            }
         } catch (IOException e) {
             // TODO: how should we handle errors?
             e.printStackTrace();
